+++ core/camel-console/src/main/java/org/apache/camel/impl/console/RouteDevConsole.java
@@ -411,7 +411,7 @@ public class RouteDevConsole extends AbstractDevConsole {
         if (percent) {
             double p;
             if (total > 0) {
-                p = (covered / total) * 100;
+                p = (covered /(double) total) * 100;
             } else {
                 p = 0;
             }
